+++ inferno-core/app/Main.hs
@@ -1,10 +1,11 @@
 {-# LANGUAGE LambdaCase #-}
+{-# LANGUAGE NamedFieldPuns #-}
 
 module Main where
 
-import qualified Data.Map as M
+import qualified Data.Map as Map
 import qualified Data.Text.IO as Text
-import Inferno.Core (Interpreter (evalInEnv, parseAndInferTypeReps), mkInferno)
+import Inferno.Core (Interpreter (..), mkInferno)
 import Inferno.Module.Prelude (builtinModules)
 import Inferno.Utils.Prettyprinter (showPretty)
 import System.Environment (getArgs)
@@ -15,13 +16,14 @@ main :: IO ()
 main = do
   file <- head <$> getArgs
   src <- Text.readFile file
-  let inferno = mkInferno builtinModules :: Interpreter ()
-  case parseAndInferTypeReps inferno src of
+  Interpreter {evalExpr, defaultEnv, parseAndInferTypeReps} <-
+    mkInferno builtinModules :: IO (Interpreter ())
+  case parseAndInferTypeReps src of
     Left err -> do
       hPutStrLn stderr $ show err
       exitFailure
     Right ast ->
-      evalInEnv inferno M.empty M.empty ast >>= \case
+      evalExpr defaultEnv Map.empty ast >>= \case
         Left err -> do
           hPutStrLn stderr $ show err
           exitFailure