+++ inferno-core/src/Inferno/Module/Prelude.hs
@@ -1,5 +1,4 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
-{-# LANGUAGE ExplicitForAll #-}
 {-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE QuasiQuotes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
@@ -132,29 +131,28 @@ baseOpsTable moduleMap =
    in IntMap.unionWith (<>) ops (IntMap.map (\xs -> [(fix, Scope modNm, op) | (fix, _, op) <- xs]) ops)
 
 builtinModulesOpsTable :: forall m c. (MonadThrow m, Pretty c, Eq c) => ModuleMap m c -> Map.Map ModuleName OpsTable
-builtinModulesOpsTable moduleMap = Map.map (\Module {moduleOpsTable} -> moduleOpsTable) $ moduleMap
+builtinModulesOpsTable = Map.map (\Module {moduleOpsTable} -> moduleOpsTable)
 
 builtinModulesPinMap :: forall m c. (MonadThrow m, Pretty c, Eq c) => ModuleMap m c -> Map.Map (Scoped ModuleName) (Map.Map Namespace (Pinned VCObjectHash))
 builtinModulesPinMap moduleMap =
   Pinned.openModule "Base" $
     Pinned.insertBuiltinModule $
       Map.foldrWithKey Pinned.insertHardcodedModule mempty $
-        Map.map (Map.map Builtin . pinnedModuleNameToHash) $
-          moduleMap
+        Map.map (Map.map Builtin . pinnedModuleNameToHash) moduleMap
 
 builtinModulesTerms :: forall m c. (MonadThrow m, Pretty c, Eq c) => ModuleMap m c -> ImplEnvM m c (TermEnv VCObjectHash c (ImplEnvM m c))
-builtinModulesTerms moduleMap = combineTermEnvs moduleMap
+builtinModulesTerms = combineTermEnvs
 
 preludeNameToTypeMap :: forall m c. (MonadThrow m, Pretty c, Eq c) => ModuleMap m c -> Map.Map (Maybe ModuleName, Namespace) (TypeMetadata TCScheme)
 preludeNameToTypeMap moduleMap =
   let unqualifiedN2h = pinnedModuleNameToHash $ moduleMap Map.! "Base"
       n2h =
         Map.unions $
-          (Map.mapKeys (Nothing,) $ pinnedModuleNameToHash builtinModule)
-            : (Map.mapKeys (Nothing,) $ unqualifiedN2h)
-            : [Map.mapKeys (Just nm,) $ (pinnedModuleNameToHash m `Map.difference` unqualifiedN2h) | (nm, m) <- Map.toList $ moduleMap]
-      h2ty = Map.unions $ pinnedModuleHashToTy builtinModule : [pinnedModuleHashToTy m | m <- Map.elems $ moduleMap]
-   in Map.mapMaybe (\h -> Map.lookup h h2ty) n2h
+          Map.mapKeys (Nothing,) (pinnedModuleNameToHash builtinModule)
+            : Map.mapKeys (Nothing,) unqualifiedN2h
+            : [Map.mapKeys (Just nm,) (pinnedModuleNameToHash m `Map.difference` unqualifiedN2h) | (nm, m) <- Map.toList moduleMap]
+      h2ty = Map.unions $ pinnedModuleHashToTy builtinModule : [pinnedModuleHashToTy m | m <- Map.elems moduleMap]
+   in Map.mapMaybe (`Map.lookup` h2ty) n2h
 
 -- In the definitions below, ###!x### is an anti-quotation to a haskell variable `x` of type `Monad m => (Value m)`
 -- This sort of Value is necessary for polymorphic functions such as `map` or `id`