+++ inferno-ml-remote/CHANGELOG.md
@@ -1,8 +1,8 @@
-# Revision History for inferno-ml
+# Revision History for inferno-ml-remote
 *Note*: we use https://pvp.haskell.org/ (MAJOR.MAJOR.MINOR.PATCH)
 
 ## 0.1.1.0 -- 2023-09-18
 * Update to use new Interpreter API that pre-computes and shares prelude
 
-## 0.1.0 -- 2023-05-09
-* First version: adds tensor type and primitives
+## 0.1.0.0
+* Initial version